Move chart specs from the UI package into the data package (#104)
* Move code to data package
